LayoutTests/resize-observer/contain-instrinsic-size-should-not-leak-nodes.html

<!DOCTYPE html>
<html>
<style>
.size-100-50 {
    width: 100px;
    height: 50px;
}
.cis-auto {
    contain-intrinsic-size: auto 1px auto 2px;
}

.skip-contents {
    content-visibility: hidden;
}

</style>
<body>
    <pre id="log">This tests element is with CIS auto and removing the element from the document.
The element should become eligible for GC at some point.
</pre>

<div id="container"></div>

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/gc.js"></script>
<script>
let initialNodeCount = internals.referencingNodeCount(document);
const CISCount = 100;

function nextRendering() {
    return new Promise(resolve => {
        requestAnimationFrame(() => requestAnimationFrame(() => resolve()));
    });
}

function createNodes() {
    const parent = document.createElement('div');
    for (let i = 0; i < CISCount; ++i) {
        const element = document.createElement('div');
        element.className = "size-100-50 cis-auto";
        parent.appendChild(element);
    }

    container.appendChild(parent);
}

function removeNodes() {
    container.removeChild(container.firstChild);
}

function skipContent() {
    const elements = container.firstChild.children;
    for (let i = 0; i < elements.length; i++) {
        elements[i].classList.add("skip-contents");
        elements[i].classList.remove("size-100-50");
    }
}

promise_test(async function () {
    createNodes();
    await nextRendering();
    removeNodes();
    await nextRendering();
    gc();
    await nextRendering();
    assert_equals(internals.referencingNodeCount(document) < initialNodeCount + CISCount * 0.8, true, 'More than 20% of nodes should be collected');
}, "Remove nodes after observing");

promise_test(async function () {
    let container = document.getElementById("container");
    createNodes();
    await nextRendering();
    skipContent();
    await nextRendering();
    removeNodes();
    await nextRendering();
    gc();
    await nextRendering();
    assert_equals(internals.referencingNodeCount(document) < initialNodeCount + CISCount * 0.8, true, 'More than 20% of nodes should be collected');
}, "Remove nodes after skipping content");
</script>
</body>

</html>
